Fix stripe credential importer
The stripe credential importer did not account for the fact
that there might be other toml keys set in the root of the
config file, not just a list of projects. By addressing
this, credentials can be importer from stripe config files
even if there are other config options in the root of the
config file.