Remove file provisioning security feature opt-outs
This commit removes the possibility of opting out of
only allowing the current process to read the provisioned
credential file and the possibility to set the file
permissions. This is done to adhere to the current security
model and due to the low probability of needing these
opt-outs in the near future. In the future we can add
them back if needed.