Update rpc layer to conform to new way of specifiying provisioners
After updating the plugin schema to have 1 default provisioner per credential type
which can be overridden by referencing the credential by name and specifying a new
provisioner on the executable level, this commit updates the protocol used by dynamically
loaded plugins, to conform to these changes.

The main change in the protocol is that credentials are now only referenced by a single index,
as they are all defined in a list on the plugin level and provisioners are identified by a
credential id and optionally an executable id, in case they have overridden the default provisioner
of a credential for a given executable.