Allow format strings in SetPathAsArg option
The file provisioner appended the output file path to the args, but
that doesn't work everywhere. Examples of where more flexibility is
needed are flags that require an "=" sign between the key and value.
For those situations, format strings can be used to tell where the
output path should go (e.g. "--config-file=%s").