Remove pointer from ProvisionerID
Pointers cannot be used in the ProvisionerID struct because
they will not resolve to the same value between executions.

Therefore, the ProvisionerID has been rewritten to not use any
pointers. The used addressing schema has also been changed to
bring it more in line with the addresses of other structs: the
slice-indexes are used as the base of the address. This has the
added benefit that it guarantees uniqueness. The previous scheme
could lead to conflicts in case a duplicate credential name was used.