Add Cachix plugin
Adds a plugin for Cachix to enable authentication using a cachix or
personal auth token. Both types of tokens are looked up via a
`$CACHIX_AUTH_TOKEN` environment variable or a
`~/.config/cachix/cachix.dhall` configuration file. This change adds
support for the former at the moment for a minimal change.

Signed-off-by: micnncim <micnncim@gmail.com>