Add more ways to get in contact for help
We're super eager to help anyone wanting to start contributing,
I've added a link to book a pairing session with one of our
teammates and another mention of the Slack space where we're
hanging out to help.

In addition, I've added a section on community & support, where I've
added the Slack space, the newsletter and filing issues, so it's
clear for folks visiting the README how they can get involved and
stay up-to-date, if they so wish.