fix(tea): use `os.UserConfigDir` to find config
On macOS, `tea` writes its configuration to
`~/Library/Application Support/tea` and not `~/.config/tea`. This
breaks the tea plugin. Go provides `os.UserConfigDir` to abstract the
location of the config directory, so use that to fix macOS while not
breaking Linux. In case `os.UserConfigDir` fails, fall back to
`~/.config/tea` (this shouldn't happen unless something weird is going
on, like the user's $HOME is not set.)

Signed-off-by: Mark Cornick <mcornick@mcornick.com>