Add ManagementURL for Sourcegraph plugin
In the dashboard the management URL is dynamic; It contains the username:
https://sourcegraph.com/users/<username>/settings/token

That's probably why it wasn't included in the first iteration of
this plugin.

I just found that https://sourcegraph.com/user/settings/tokens redirects
to the management URL for the currently logged in user. Sourcegraph also
uses this in the 'src login' command, so it should be safe to use that
here as well.