Fix signed commits CI job
The job was previously ran on "pull_request_target". This runs the
job on the base branch (i.e. the `main` branch of the
 1password/shell-plugins repo). Since we want to know if commits
are signed on all the commits to be merged (i.e. also on the fork),
I have changed this job to run on the merge commit instead.

Relevant documentation:
https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull_request_target

Note that the usage example of https://github.com/1Password/check-signed-commits-action also recommends running on the "pull_reqeusts" event.