Add needsauth specification
Add a list of commands that don't require authentication. When running
those commands with the cli, the plugin will not try to fetch the token
from the vault.

I opted for a negative list of commands over a positive one since the
latter would end up being much longer, would need to be more complete,
and would probably need more updates with time, with newer npm versions.

<!-- ps-id: ec70313e-76dd-4aab-811d-196348b499ac -->