Update terraform.go
The 1Password Terraform shell plugin successfully provides credentials for commands like `terraform apply` and `terraform init`, but it specifically fails to provide them for the `terraform output` command. A key factor is the use of a remote state file in S3, which requires authentication. The issue is not present when credentials are set directly as environment variables.

As an example, the `output` subcommand is missing from this list. I suspect this causes the plugin to omit credential injection for the `output` command. As a result, the terraform process fails to receive the credentials in its environment. Instead terraform falls back to the configuration in `provider.tf` file. Since that file doesn't contain explicit credentials, the backend authentication fails.